Subtract shipping costs from commissionable amount
Previously, we sent the entire "Grand Total" to AWIN as the order value
which should be used for commission. This was incorrect, as AWIN wishes
us to exclude the shipping amount from the total. AWIN have been
informed of the error in their Integration Document, which called for
the "Total Transaction Amount".