Subtract taxes from commissionable amount
The Integration Document sent by AWIN contained an error, calling for
inclusion of VAT in the totals sent to AWIN. Further discussion with
AWIN confirmed this was an error, but not prior to the initial release
of this module.

There are potentially some edge-cases surrounding how discounts and taxes
are calculated. Specifically: Magento is highly configurable in the way
it determines when taxes and discounts are applied. As a result, there
are some configurations where excluding the "shipping amount" may not
actually exclude the full cost of shipping, as taxes may be applied
separately. This situation can be exacerbated by discounts to shipping
costs, which may interact with tax configurations in unexpected ways,
resulting in cases in which the "reduced" shipping amount is excluded,
but the taxes on the shipping amount are not adjusted. In such
situations, the commissionable value will almost certainly be wrong.
These cases are explicitly not handled in this version, and the README
has been updated to include a note regarding this.

Adding support for additional tax configurations should be relatively
simple, but won't be attempted without a concrete example to use for
reference. Perhaps someone will open an issue in the future to highlight
a different configuration.